Implementation of Hateoas links in the phones.
