+++ src/Application/APIs/Phones/All/Helpers/Loader.php
@@ -3,11 +3,12 @@
 namespace App\Application\APIs\Phones\All\Helpers;
 
 use App\Application\APIs\Exceptions\ItemNotFoundException;
+use App\Application\APIs\Helpers\Hateoas\Interfaces\HateoasBuilderInterface;
 use App\Application\APIs\Interfaces\InputInterface;
 use App\Application\APIs\Interfaces\OutputListInterface;
 use App\Application\APIs\Phones\All\Helpers\Interfaces\LoaderInterface;
+use App\Application\APIs\Phones\All\OutputItems\PhoneOutput;
 use App\Application\APIs\Phones\All\OutputList\PhonesOutputList;
-use App\Domain\OutputItems\PhoneOutput;
 use App\Domain\Repositories\PhoneRepository;
 
 class Loader implements LoaderInterface
@@ -16,15 +17,22 @@ class Loader implements LoaderInterface
      * @var PhoneRepository
      */
     private $phoneRepository;
+    /**
+     * @var HateoasBuilderInterface
+     */
+    private $hateoasBuilder;
 
     /**
      * Loader constructor.
      * @param PhoneRepository $phoneRepository
+     * @param HateoasBuilderInterface $hateoasBuilder
      */
     public function __construct(
-        PhoneRepository $phoneRepository
+        PhoneRepository $phoneRepository,
+        HateoasBuilderInterface $hateoasBuilder
     ) {
         $this->phoneRepository = $phoneRepository;
+        $this->hateoasBuilder = $hateoasBuilder;
     }
 
     /**
@@ -57,7 +65,10 @@ class Loader implements LoaderInterface
         $list = [];
 
         foreach ($phones as $phone) {
-            $list[] = new PhoneOutput($phone, []);
+            $list[] = new PhoneOutput(
+                $phone,
+                $this->hateoasBuilder->build('Phone_show', ['id' => $phone['id']], 'self', 'GET')
+            );
         }
 
         return new PhonesOutputList($list);
